requirements.txt
pypdf
langchain
langchain_community
langchain_openai
chromadb # Vector storage
pytest
boto3
streamlit
python-docx
python-pptx
